requirements.txt
# Main dependencies
flask==2.3.3
requests==2.31.0
anthropic==0.51.0
python-dotenv==1.0.0

# Testing dependencies
pytest==8.1.1
pytest-cov==4.1.0
pytest-mock==3.14.0
